# api/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "oms",
    "version": "last"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/v1/host": {
      "put": {
        "tags": [
          "host"
        ],
        "summary": "修改host",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "required": true,
            "description": "主机ID",
            "type": "string"
          },
          {
            "name": "hostname",
            "in": "formData",
            "required": false,
            "description": "主机名",
            "type": "string"
          },
          {
            "name": "user",
            "in": "formData",
            "required": false,
            "description": "用户名",
            "type": "string"
          },
          {
            "name": "addr",
            "in": "formData",
            "required": false,
            "description": "主机地址",
            "type": "string"
          },
          {
            "name": "port",
            "in": "formData",
            "required": false,
            "description": "ssh端口(默认：22)",
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": false,
            "description": "ssh密码",
            "type": "string"
          },
          {
            "name": "group",
            "in": "formData",
            "required": false,
            "description": "主机组ID",
            "type": "string"
          },
          {
            "name": "keyFile",
            "in": "formData",
            "required": false,
            "description": "秘钥文件",
            "type": "file"
          },
          {
            "name": "tags",
            "in": "formData",
            "required": false,
            "description": "主机标签集合",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "user": {
                      "type": "string"
                    },
                    "addr": {
                      "type": "string"
                    },
                    "port": {
                      "type": "number"
                    },
                    "status": {
                      "type": "boolean"
                    },
                    "group_id": {
                      "type": "number"
                    },
                    "group": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "mode": {
                          "type": "number"
                        },
                        "params": {
                          "type": "string"
                        }
                      }
                    },
                    "tags": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "host"
        ],
        "summary": "新增host",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "hostname",
            "in": "formData",
            "required": true,
            "description": "主机名",
            "type": "string"
          },
          {
            "name": "user",
            "in": "formData",
            "required": true,
            "description": "用户名",
            "type": "string"
          },
          {
            "name": "addr",
            "in": "formData",
            "required": true,
            "description": "主机地址",
            "type": "string"
          },
          {
            "name": "port",
            "in": "formData",
            "required": true,
            "description": "ssh端口(默认：22)",
            "type": "string"
          },
          {
            "name": "group",
            "in": "formData",
            "required": true,
            "description": "主机组ID",
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "description": "ssh密码",
            "type": "string"
          },
          {
            "name": "keyFile",
            "in": "formData",
            "required": false,
            "description": "ssh秘钥",
            "type": "file"
          },
          {
            "name": "tags",
            "in": "formData",
            "required": false,
            "description": "tag的id集合",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "user": {
                      "type": "string"
                    },
                    "addr": {
                      "type": "string"
                    },
                    "port": {
                      "type": "number"
                    },
                    "status": {
                      "type": "boolean"
                    },
                    "group_id": {
                      "type": "number"
                    },
                    "group": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "mode": {
                          "type": "number"
                        },
                        "params": {
                          "type": "string"
                        }
                      }
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "host"
        ],
        "summary": "获取所有host",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "user": {
                        "type": "string"
                      },
                      "addr": {
                        "type": "string"
                      },
                      "port": {
                        "type": "number"
                      },
                      "status": {
                        "type": "boolean"
                      },
                      "private_key_id": {
                        "type": "number"
                      },
                      "group_id": {
                        "type": "number"
                      },
                      "group": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "mode": {
                            "type": "number"
                          },
                          "params": {
                            "type": "string"
                          }
                        }
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ]
                        }
                      },
                      "tunnels": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "mode": {
                              "type": "string"
                            },
                            "source": {
                              "type": "string"
                            },
                            "destination": {
                              "type": "string"
                            },
                            "status": {
                              "type": "number"
                            },
                            "error_msg": {
                              "type": "string"
                            },
                            "host_id": {
                              "type": "number"
                            }
                          }
                        }
                      },
                      "jobs": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            },
                            "type": {
                              "type": "string"
                            },
                            "spec": {
                              "type": "string"
                            },
                            "cmd": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string"
                            },
                            "host_id": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "type",
                            "spec",
                            "cmd",
                            "status",
                            "host_id"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/host/:id": {
      "delete": {
        "tags": [
          "host"
        ],
        "summary": "删除host",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "主机ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "null"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "host"
        ],
        "summary": "通过ID获取host",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "主机的ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "user": {
                      "type": "string"
                    },
                    "addr": {
                      "type": "string"
                    },
                    "port": {
                      "type": "number"
                    },
                    "status": {
                      "type": "boolean"
                    },
                    "private_key_id": {
                      "type": "number"
                    },
                    "group_id": {
                      "type": "number"
                    },
                    "group": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "mode": {
                          "type": "number"
                        },
                        "params": {
                          "type": "string"
                        }
                      }
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name"
                        ]
                      }
                    },
                    "tunnels": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "mode": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          },
                          "destination": {
                            "type": "string"
                          },
                          "status": {
                            "type": "number"
                          },
                          "error_msg": {
                            "type": "string"
                          },
                          "host_id": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "jobs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "spec": {
                            "type": "string"
                          },
                          "cmd": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "host_id": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "type",
                          "spec",
                          "cmd",
                          "status",
                          "host_id"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tag": {
      "put": {
        "tags": [
          "tag"
        ],
        "summary": "修改tag",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "required": true,
            "description": "标签ID",
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "required": false,
            "description": "标签名",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tag"
        ],
        "summary": "新增tag",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "description": "标签名",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tag"
        ],
        "summary": "获取所有tag",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tag/:id": {
      "delete": {
        "tags": [
          "tag"
        ],
        "summary": "删除tag",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "标签ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "null"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tag"
        ],
        "summary": "通过ID获取tag",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "标签ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/group/:id": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "通过ID获取group",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "组ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "number"
                    },
                    "params": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "group"
        ],
        "summary": "删除group",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "组ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "data": {
                  "type": "null"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tools/download": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "下载文件",
        "description": "返回文件流",
        "parameters": [
          {
            "name": "host_id",
            "in": "query",
            "required": true,
            "description": "host的ID",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "文件路径",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {}
            }
          }
        }
      }
    },
...
    "/tools/import": {
      "post": {
        "tags": [
          "other"
        ],
        "summary": "数据库导入",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "dataFile",
            "in": "formData",
            "required": true,
            "description": "导入的文件",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "title": "empty object",
              "properties": {}
            }
          }
        }
      }
    },
...
  }
}